+++ lib/app/injection/injection.dart
@@ -15,7 +15,7 @@ import 'package:givt_app/features/children/details/repositories/child_details_re
 import 'package:givt_app/features/children/edit_child/repositories/create_child_repository.dart';
 import 'package:givt_app/features/children/edit_profile/repositories/edit_profile_repository.dart';
 import 'package:givt_app/features/children/family_goal/repositories/create_family_goal_repository.dart';
-import 'package:givt_app/features/children/family_goal_tracker/repository/goal_tracker_repository.dart';
+import 'package:givt_app/features/children/goal_tracker/repository/goal_tracker_repository.dart';
 import 'package:givt_app/features/children/family_history/repository/family_history_repository.dart';
 import 'package:givt_app/features/children/generosity_challenge/repositories/generosity_challenge_repository.dart';
 import 'package:givt_app/features/children/overview/repositories/family_overview_repository.dart';
