+++ lib/features/children/goal_tracker/widgets/goal_active_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:givt_app/features/children/family_goal_tracker/cubit/goal_tracker_cubit.dart';
-import 'package:givt_app/features/children/family_goal_tracker/widgets/gradient_progress_bar.dart';
+import 'package:givt_app/features/children/goal_tracker/cubit/goal_tracker_cubit.dart';
+import 'package:givt_app/features/children/goal_tracker/widgets/gradient_progress_bar.dart';
 import 'package:givt_app/utils/app_theme.dart';
 
 class GoalActiveWidget extends StatelessWidget {
