+++ lib/features/children/goal_tracker/widgets/gradient_progress_bar.dart
