+++ lib/features/children/overview/widgets/family_available_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:givt_app/features/auth/cubit/auth_cubit.dart';
-import 'package:givt_app/features/children/family_goal_tracker/pages/family_goal_tracker.dart';
+import 'package:givt_app/features/children/goal_tracker/pages/family_goal_tracker.dart';
 import 'package:givt_app/features/children/family_history/family_history.dart';
 import 'package:givt_app/features/children/family_history/family_history_cubit/family_history_cubit.dart';
 import 'package:givt_app/features/children/overview/cubit/family_overview_cubit.dart';
